// csm/csp-angular19/src/app/core/models/audit-checklist-based-model.ts
export class ServiceAreaModelNew {
    id: number = 0;
    title: string = '';
    description: string = '';
    createD_BY: string = localStorage.getItem('empid') || '';
    createD_DATE: Date = new Date();
    updateD_BY: string = localStorage.getItem('empid') || '';
    updateD_DATE: Date = new Date();
    isactive: boolean = true;
    isMapped: boolean = false;
    show_in_Master: boolean = true;
}

export class ProcessAreaModelNew {
    id: number = 0;
    title: string = '';
    description: string = '';
    createD_BY: string = localStorage.getItem('empid') || '';
    createD_DATE: Date = new Date();
    updateD_BY: string = localStorage.getItem('empid') || '';
    updateD_DATE: Date = new Date();
    isactive: boolean = true;
    checked: boolean = false;
    show_in_Master: boolean = true;
}

export class ProcessModelProcessMapping {
    id: number = 0;
    procesS_MODEL_ID: number = 0;
    procesS_ID: number = 0;
    createD_BY: string = localStorage.getItem('empid') || '';
    createD_DATE: Date = new Date();
    updateD_BY: string = localStorage.getItem('empid') || '';
    updateD_DATE: Date = new Date();
    isactive: boolean = true;
}

export class ProcessServiceAreaMapping {
    id: number = 0;
    servicE_AREA_ID: number = 0;
    procesS_ID: number = 0;
    createD_BY: string = localStorage.getItem('empid') || '';
    createD_DATE: Date = new Date();
    updateD_BY: string = localStorage.getItem('empid') || '';
    updateD_DATE: Date = new Date();
    isactive: boolean = true;
}
// ...
export class ProcessModelNew {
    id: number = 0;
    procesS_AREA_ID: number = 0;
    title: string = "";
    description: string = "";
    procesS_MODEL_REFERENCE_LIST: number[] = [];
    createD_BY: string = localStorage.getItem('empid') || '';
    createD_DATE: Date = new Date();
    updateD_BY: string = localStorage.getItem('empid') || '';
    updateD_DATE: Date = new Date();
    isactive: boolean = true;
    bSelected: boolean = false;
    show_in_Master: boolean = true;
}

export class ProcessServiceAreaMappingList {
    id: number = 0;
    serviceAreaId: number = 0;
    processAreaId: number = 0;
    processTitle: string = '';
    processDescription: string = '';
    serviceAreaName: string = '';
    processArea: string = '';
    processId: number = 0;
    createD_BY: string = localStorage.getItem('empid') || '';
    createD_DATE: Date = new Date();
    updateD_BY: string = localStorage.getItem('empid') || '';
    updateD_DATE: Date = new Date();
    isactive: boolean = true;
}